Dto пренесены в Domain
