+++ EmploymentAgency/EmploymentAgency.Domain/DTO/ResumeDto.cs
@@ -1,4 +1,4 @@
-﻿namespace EmploymentAgency.Server.DTO;
+﻿namespace EmploymentAgency.Domain.DTO;
 
 /// <summary>
 /// DTO для класса Резюме
